feat: replace authentication with auth
