+++ expediagroup-sdk-core/src/main/kotlin/com/expediagroup/sdk/core/auth/oauth/OAuthTokenResponse.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.expediagroup.sdk.core.authentication.oauth
+package com.expediagroup.sdk.core.auth.oauth
 
 import com.expediagroup.sdk.core.common.getOrThrow
 import com.expediagroup.sdk.core.exception.client.ExpediaGroupResponseParsingException
