+++ expediagroup-sdk-core/src/test/kotlin/com/expediagroup/sdk/core/auth/common/AuthUtilsTest.kt
@@ -1,4 +1,4 @@
-package com.expediagroup.sdk.core.authentication.common
+package com.expediagroup.sdk.core.auth.common
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
