+++ expediagroup-sdk-core/src/test/kotlin/com/expediagroup/sdk/core/auth/oauth/OAuthCredentialsTest.kt
@@ -1,4 +1,4 @@
-package com.expediagroup.sdk.core.authentication.oauth
+package com.expediagroup.sdk.core.auth.oauth
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
