+++ expediagroup-sdk-core/src/test/kotlin/com/expediagroup/sdk/core/auth/oauth/OAuthManagerTest.kt
@@ -1,4 +1,4 @@
-package com.expediagroup.sdk.core.authentication.oauth
+package com.expediagroup.sdk.core.auth.oauth
 
 import com.expediagroup.sdk.core.exception.client.ExpediaGroupResponseParsingException
 import com.expediagroup.sdk.core.exception.client.ExpediaGroupTransportException
@@ -27,17 +27,17 @@ import org.junit.jupiter.api.assertThrows
 import java.util.UUID
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class OAuthAuthenticationManagerTest {
+class OAuthManagerTest {
     private lateinit var transport: Transport
     private lateinit var credentials: OAuthCredentials
-    private lateinit var authenticationManager: OAuthAuthenticationManager
+    private lateinit var authenticationManager: OAuthManager
     private val authUrl = "https://auth.example.com/token"
 
     @BeforeAll
     fun setup() {
         transport = mockk()
         credentials = OAuthCredentials("client_key", "client_secret")
-        authenticationManager = OAuthAuthenticationManager(authUrl, credentials, transport)
+        authenticationManager = OAuthManager(authUrl, credentials, transport)
     }
 
     @AfterEach
